requirements.txt
python-pptx~=0.6.23
pillow~=10.3.0